lambda/requirements.txt
aws-cdk-lib==2.42.0
constructs>=10.0.0,<11.0.0

boto3~=1.23.6
requests~=2.28.1
botocore~=1.26.10